requirements.txt
# filepath: e:\RDC CONCRETE PROJECT\requirements.txt
flask==3.0.2
flask-wtf==1.2.1 # If using Flask-WTF forms elsewhere
werkzeug==3.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
numpy~=1.26 # Or 2.x if compatible with other libs
opencv-python~=4.9 # Or latest 4.x
Pillow~=10.2 # Or latest 10.x
ultralytics~=8.1 # Or latest 8.x (includes PyTorch)
pandas==2.2.4
scikit-image==0.25.0
scikit-learn==1.5.1
scipy==1.15.0
seaborn==0.13.2